+++ SourceCode/Data/ModuleExit.cs
@@ -6,11 +6,11 @@ namespace ModulesRegistry.Data
     {
         public int Id { get; set; }
         public int ModuleId { get; set; }
-        public int GablePropertyId { get; set; }
+        public int GableTypeId { get; set; }
         public string Label { get; set; }
         public int Direction { get; set; }
 
         public virtual Module Module { get; set; }
-        public virtual Property GableProperty { get; set; }
+        public virtual ModuleGableType GableType { get; set; }
     }
 }